restaurante_qr_project/requirements.txt
# ====================================
# SGIR - Sistema de Gestión Integral Restaurante
# Dependencias Python
# ====================================

# Framework principal
Django==5.1.4

# Django REST Framework y JWT
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

# CORS y seguridad
django-cors-headers==4.6.0

# Archivos estáticos en producción
whitenoise==6.8.2

# Servidor WSGI para producción
gunicorn==23.0.0

# Variables de entorno
python-decouple==3.8

# Base de datos PostgreSQL
psycopg2-binary==2.9.10

# Generación de códigos QR
qrcode==8.0
Pillow==11.0.0

# Filtros para APIs
django-filter==24.3

# Panel de administración
django-admin-interface==0.28.9
django-colorfield==0.11.0

# Reportes Excel y PDF
openpyxl==3.1.5
reportlab==4.2.5

# Testing y calidad
pytest==8.3.4
pytest-django==4.9.0

# Django extensions (management commands)
django-extensions==3.2.3
